#5 Implemented TokenFactory for creating instance of JWT token
