+++ Cooper/Controllers/AuthController.cs
@@ -5,13 +5,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
-using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
 using Cooper.Models;
 using Cooper.Controllers.ViewModels;
 using Cooper.Repository;
 using Cooper.Configuration;
 using Cooper.Services;
+using Cooper.Services.Authorization;
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace Cooper.Controllers
@@ -27,31 +26,18 @@ namespace Cooper.Controllers
         }
 
         [HttpPost, Route("login")]
-        public IActionResult Login([FromBody]Login user)
+        public IActionResult Login([FromBody]Login login)
         {
-            if (user == null)
+            if (login == null)
             {
                 return BadRequest("Invalid client request");
             }
 
-            bool authValid = userRepository.CheckCredentials(user.Username, user.Password);
+            bool authValid = userRepository.CheckCredentials(login.Username, login.Password);
 
             if (authValid)
             {
-
-                var signinCredentials = new SigningCredentials(AuthOptions.GetSymmetricSecurityKey(), SecurityAlgorithms.HmacSha256);
-
-
-                var identity = GetIdentity(user.Username);
-
-                var jwt = new JwtSecurityToken(
-                        issuer: AuthOptions.ISSUER,
-                        audience: AuthOptions.AUDIENCE,
-                        claims: identity.Claims,
-                        expires: DateTime.Now.AddMinutes(AuthOptions.LIFETIME),
-                        signingCredentials: signinCredentials);
-
-                string tokenString = new JwtSecurityTokenHandler().WriteToken(jwt);
+                string tokenString = new TokenFactory(login).GetTokenString();
                 
                 return Ok(new { Token = tokenString });
             }
@@ -61,18 +47,6 @@ namespace Cooper.Controllers
             }
         }
 
-        private ClaimsIdentity GetIdentity(string username)
-        {
-            var claims = new List<Claim>
-                {
-                    new Claim("username", username)
-                };
-
-            ClaimsIdentity claimsIdentity =
-            new ClaimsIdentity(claims, "Token", ClaimsIdentity.DefaultNameClaimType,
-                ClaimsIdentity.DefaultRoleClaimType);
-
-            return claimsIdentity;
-        }
+        
     }
 }
